+++ mayan/apps/common/literals.py
@@ -14,14 +14,8 @@ DEFAULT_COMMON_HOME_VIEW_DASHBOARD_NAME = 'user'
 DEFAULT_COMMON_PROJECT_TITLE = mayan.__title__
 DEFAULT_COMMON_PROJECT_URL = mayan.__website__
 
-DJANGO_SQLITE_BACKEND = 'django.db.backends.sqlite3'
-
 MESSAGE_DEPRECATION_WARNING = _(
     'This feature has been deprecated and will be removed in a future version.'
-)
-MESSAGE_SQLITE_WARNING = _(
-    'Your database backend is set to use SQLite. SQLite should only be used '
-    'for development and testing, not for production.'
 )
 
 TIME_DELTA_UNIT_DAYS = 'days'