+++ mayan/apps/databases/utils.py
@@ -1,4 +1,11 @@
+from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
+
+from .literals import DJANGO_SQLITE_BACKEND
+
+
+def check_for_sqlite():
+    return settings.DATABASES['default']['ENGINE'] == DJANGO_SQLITE_BACKEND and settings.DEBUG is False
 
 
 def check_queryset(self, queryset):